requirements.txt
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
slack-sdk
slack-bolt
Flask
gunicorn
gensim
sumy
openpyxl
pytube
youtube_dl
google-cloud-speech
textblob 
googletrans
matplotlib